address comments and create registries
