+++ xds/src/main/java/io/grpc/xds/internal/matcher/MatcherRegistry.java
@@ -1,0 +1,49 @@
+/*
+ * Copyright 2026 The gRPC Authors
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package io.grpc.xds.internal.matcher;
+
+import com.google.common.annotations.VisibleForTesting;
+import java.util.Map;
+import java.util.concurrent.ConcurrentHashMap;
+import javax.annotation.Nullable;
+
+/**
+ * Registry for {@link MatcherProvider}.
+ */
+public final class MatcherRegistry {
+  private static final MatcherRegistry DEFAULT_INSTANCE = new MatcherRegistry();
+
+  private final Map<String, MatcherProvider> matcherProviders = new ConcurrentHashMap<>();
+
+  public static MatcherRegistry getDefaultRegistry() {
+    return DEFAULT_INSTANCE;
+  }
+
+  @VisibleForTesting
+  public MatcherRegistry() {
+    register(new CelStateMatcher.Provider());
+  }
+
+  public void register(MatcherProvider provider) {
+    matcherProviders.put(provider.typeUrl(), provider);
+  }
+
+  @Nullable
+  public MatcherProvider getMatcherProvider(String typeUrl) {
+    return matcherProviders.get(typeUrl);
+  }
+}
